Check if adaPot or gov module is enabled before adapot/gov processing. This is required for GraalVM native image
