+++ aggregates/adapot/src/main/java/com/bloxbean/cardano/yaci/store/adapot/AdaPotProperties.java
@@ -11,6 +11,8 @@ import lombok.NoArgsConstructor;
 @Builder
 public class AdaPotProperties {
 
+    private boolean enabled;
+
     @Builder.Default
     private int updateRewardDbBatchSize = 200;
