+++ aggregates/adapot/src/main/java/com/bloxbean/cardano/yaci/store/adapot/processor/GenesisPoolProcessor.java
@@ -7,6 +7,7 @@ import com.bloxbean.cardano.client.common.model.Networks;
 import com.bloxbean.cardano.yaci.core.model.CredentialType;
 import com.bloxbean.cardano.yaci.core.model.Era;
 import com.bloxbean.cardano.yaci.core.util.HexUtil;
+import com.bloxbean.cardano.yaci.store.adapot.AdaPotProperties;
 import com.bloxbean.cardano.yaci.store.adapot.domain.EpochStake;
 import com.bloxbean.cardano.yaci.store.adapot.service.AdaPotService;
 import com.bloxbean.cardano.yaci.store.common.config.StoreProperties;
@@ -43,6 +44,7 @@ import static com.bloxbean.cardano.yaci.store.adapot.jooq.Tables.EPOCH_STAKE;
 @Slf4j
 public class GenesisPoolProcessor {
     private final StoreProperties storeProperties;
+    private final AdaPotProperties adaPotProperties;
     private final PoolStorage poolStorage;
     private final PoolCertificateStorage poolCertificateStorage;
     private final StakingCertificateStorage stakingCertificateStorage;
@@ -53,6 +55,9 @@ public class GenesisPoolProcessor {
     @EventListener
     @Transactional
     public void handleGenesisPoolRegistration(GenesisBlockEvent genesisBlockEvent) {
+        if (!adaPotProperties.isEnabled())
+            return;
+
         //If genesis block starts with Byron, don't do anything
         if (genesisBlockEvent.getEra().getValue() < Era.Shelley.getValue())
             return;
